+++ packages/docusaurus/src/ScalarDocusaurus.tsx
@@ -1,5 +1,6 @@
 import BrowserOnly from '@docusaurus/BrowserOnly'
 import { type ReferenceProps } from '@scalar/api-reference-react'
+import '@scalar/api-reference-react/style.css'
 import Layout from '@theme/Layout'
 import React from 'react'