+++ packages/components/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @scalar/components
 
+## 0.9.0
+
+### Minor Changes
+
+- 7d3bc16: feat: css files must be explicitly imported, check the docs
+- 7d3bc16: BREAKING CHANGE: css must now be explicitly imported, see docs for more details
+
+## 0.8.0
+
+### Minor Changes
+
+- 18ecd33: refactor: remove component app install
+- 26e2392: refactor: remove :where prefix in favor of cascade layers
+
+## 0.7.15
+
+### Patch Changes
+
+- Updated dependencies [cc5402c]
+  - @scalar/oas-utils@0.1.16
+
 ## 0.7.14
 
 ### Patch Changes