+++ packages/docusaurus/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @scalar/docusaurus
 
+## 0.3.0
+
+### Minor Changes
+
+- 7d3bc16: feat: css files must be explicitly imported, check the docs
+
+### Patch Changes
+
+- Updated dependencies [7d3bc16]
+  - @scalar/api-reference-react@0.2.0
+
+## 0.2.56
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.98
+
+## 0.2.55
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.97
+
+## 0.2.54
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.96
+
 ## 0.2.53
 
 ### Patch Changes